arranging all my js code
clean, not repeat, structure
