+++ profile/index.html
@@ -76,7 +76,7 @@
         </div>
     </div>
 <footer class="text-center mt-5">
-    <p>&copy; 2023 Socilly. All rights reserved.</p>
+    <p>&copy; 2023 Socially. All rights reserved.</p>
 </footer>
 <script src="/js/bootstrap.bundle.min.js"></script>
 </body>
